more ui updates
