--- Assets/ObjectInteraction.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: faf26151c30f6e7d9ae98c0e9b4aca8a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
