--- Assets/PickUpObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 68f7af2675b8e6c46958f38e2c8779fe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
