+++ Assets/Prefabs/UI/UIElements.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 83efc3612fb2c701a8c331b5028db9b6
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
