+++ Assets/Scenes/StartMenu.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2ce392c4d2685f83ab5fc74e35a4a137
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
