+++ Assets/Script/HUDScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 47c218fea3c5aa60ba1dfd5a10285290
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
