+++ Assets/Script/HedgehogScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ee21f12382b3c9adca9f808046ee3e8b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
