+++ Assets/Script/SaveHedgehog.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6a8763393a9979d399d78adf21b8676a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
