+++ Assets/Script/Taxi.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2a178ba53f7415995b3159cfb4bbc27a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
