+++ Assets/Script/UpdateForkliftColission.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f20baa555b3b40b19ba71ae0973a39be
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
