--- Assets/SquashObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 523f477c27c5136c296ee6d0c7c48006
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
