+++ test/test_Data/boot.config
@@ -1,8 +1,8 @@
 player-connection-mode=Listen
-player-connection-guid=1842634427
+player-connection-guid=90200434
 player-connection-debug=0
 player-connection-project-name=hitfox
-player-connection-ip=172.31.17.16
+player-connection-ip=192.168.0.87
 gfx-enable-gfx-jobs=1
 gfx-enable-native-gfx-jobs=1
 wait-for-native-debugger=0
